Add authorization check to delete mapping

+++ backend/src/main/java/de/mcstangl/projectplanner/controller/MilestoneController.java
@@ -21,7 +21,7 @@ import static org.springframework.util.Assert.hasText;
 @CrossOrigin
 @RestController
 @RequestMapping("api/project-planner/milestone")
-public class MilestoneController extends Mapper{
+public class MilestoneController extends Mapper {
 
     private final MilestoneService mileStoneService;
     private final ProjectService projectService;
@@ -55,7 +55,7 @@ public class MilestoneController extends Mapper{
     @PutMapping
     public ResponseEntity<MilestoneDto> updateMilestone(@AuthenticationPrincipal UserEntity authUser, @RequestBody MilestoneDto milestoneDto) {
 
-        if(milestoneDto.getId() == null) {
+        if (milestoneDto.getId() == null) {
             throw new IllegalArgumentException("Milestone ohne ID kann nicht upgedated werden");
         }
 
@@ -70,11 +70,13 @@ public class MilestoneController extends Mapper{
     }
 
     @DeleteMapping("{id}")
-    public ResponseEntity<MilestoneDto> deleteMilestone(@AuthenticationPrincipal  UserEntity authUser, @PathVariable Long id){
+    public ResponseEntity<MilestoneDto> deleteMilestone(@AuthenticationPrincipal UserEntity authUser, @PathVariable Long id) {
+        if (isAdmin(authUser)) {
+            MilestoneEntity milestoneEntity = mileStoneService.deleteById(id);
 
-        MilestoneEntity milestoneEntity = mileStoneService.deleteById(id);
-
-        return ok(mapMilestone(milestoneEntity));
+            return ok(mapMilestone(milestoneEntity));
+        }
+        return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
     }
 
     private MilestoneEntity getMilestoneEntity(MilestoneDto milestoneDto) {
